JS splitted and cleaned
